refactored ViewModels; moved from view to interface_adapter
