+++ src/interface_adapter/ResultViewModel.java
@@ -1,0 +1,2 @@
+package interface_adapter;public class ResultViewModel {
+}
