--- src/view/MainMenuViewModel.java
@@ -1,12 +1,0 @@
-package src.view;
-
-public class MainMenuViewModel extends ViewModel {
-    public MainMenuViewModel(String viewName) {
-        super(viewName);
-    }
-
-    @Override
-    public void propertyChanged() {
-
-    }
-}
