--- src/view/ResultViewModel.java
@@ -1,2 +1,0 @@
-package src.view;public class ResultViewModel {
-}
