--- src/view/ViewModel.java
@@ -1,12 +1,0 @@
-package src.view;
-
-public abstract class ViewModel {
-    private final String viewName;
-
-    public ViewModel(String viewName) {
-        this.viewName = viewName;
-    }
-    public String getViewName() {return viewName;}
-
-    public abstract void propertyChanged();
-}
